+++ packages/core/CHANGELOG.md
@@ -6,6 +6,10 @@
 
   Requires a policy decision point server running Cerbos 0.26+.
 
+- [`Client.enablePolicy`](../../docs/core.client.enablepolicy.md) and [`Client.enablePolicies`](../../docs/core.client.enablepolicies.md) methods to enable policies ([#475](https://github.com/cerbos/cerbos-sdk-javascript/pull/475))
+
+  Requires a policy decision point server running Cerbos 0.26+.
+
 ## [0.9.1] - 2023-03-23
 
 ### Removed